Auth and user info logic separated
