+++ app.js
@@ -7,6 +7,7 @@ import swaggerUi from "swagger-ui-express";
 import swaggerDocument from "./swagger.json" assert {"type": "json"};
 
 import authRouter from "./routes/authRoutes.js";
+import userRouter from "./routes/userRoutes.js";
 
 const { DB_HOST, PORT } = process.env;
 
@@ -19,6 +20,7 @@ app.use(express.json());
 
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 app.use("/auth", authRouter);
+app.use("/users", userRouter);
 
 app.use((_, res) => {
   res.status(404).json({ message: "Route not found" });
