+++ routes/authRoutes.js
@@ -1,14 +1,8 @@
 import express from "express";
 import authControllers from "../controllers/authControllers.js";
 import validateBody from "../middlewares/validateBody.js";
-import {
-  updateUserProfileSchema,
-  updateUserThemeSchema,
-  userLoginSchema,
-  userRegistrationSchema,
-} from "../schemas/joiSchemas/userSchemas.js";
 import authenticate from "../middlewares/authenticate.js";
-import upload from "../middlewares/upload.js";
+import { userLoginSchema, userRegistrationSchema } from "../schemas/joiSchemas/userSchemas.js";
 
 const authRouter = express.Router();
 
@@ -19,27 +13,6 @@ authRouter.post(
 );
 
 authRouter.post("/login", validateBody(userLoginSchema), authControllers.login);
-
-authRouter.patch(
-  "/",
-  authenticate,
-  validateBody(updateUserThemeSchema),
-  authControllers.updateTheme
-);
-
-authRouter.patch(
-  "/avatars",
-  authenticate,
-  upload.single("avatar"),
-  authControllers.updateAvatar
-);
-
-authRouter.put(
-  "/",
-  authenticate,
-  validateBody(updateUserProfileSchema),
-  authControllers.updateProfile
-);
 
 authRouter.get("/current", authenticate, authControllers.getCurrentUser);
